Remove implementations of PluginInstallers
Since they each only provide one tool-specific String method, it has been moved to the corresponding tool.
