+++ cli/src/main/java/com/devonfw/tools/ide/tool/androidstudio/AndroidStudio.java
@@ -9,6 +9,7 @@ import com.devonfw.tools.ide.environment.EnvironmentVariables;
 import com.devonfw.tools.ide.environment.EnvironmentVariablesType;
 import com.devonfw.tools.ide.process.ProcessMode;
 import com.devonfw.tools.ide.tool.ide.IdeToolCommandlet;
+import com.devonfw.tools.ide.tool.ide.IdeaBasedPluginInstaller;
 import com.devonfw.tools.ide.tool.ide.PluginDescriptor;
 import com.devonfw.tools.ide.version.VersionIdentifier;
 
@@ -67,14 +68,13 @@ public class AndroidStudio extends IdeToolCommandlet {
   @Override
   public void installPlugin(PluginDescriptor plugin) {
 
-    AndroidStudioPluginInstaller pluginInstaller = this.getPluginInstaller();
+    IdeaBasedPluginInstaller pluginInstaller = new IdeaBasedPluginInstaller(context, this);
     String downloadUrl = pluginInstaller.getDownloadUrl(plugin);
     pluginInstaller.installPlugin(plugin, downloadUrl);
   }
 
   @Override
-  public AndroidStudioPluginInstaller getPluginInstaller() {
-    return new AndroidStudioPluginInstaller(context, this);
+  public String getMacToolApp() {
+    return "Android Studio Preview.app";
   }
-
 }
